% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\title{Pre-Processing of models}
\usage{
pre_process(data, operations = NULL, control = NULL)
}
\arguments{
\item{data}{input data.frame}

\item{operations}{string vector with the operation names in order
to be applied}

\item{control}{list of control parameters for the operations}
}
\description{
Pre-Processing of models
}
\examples{
fn <- binh
model <- build_surmodel(fn, 20, 2)
data <- model@data
}
