% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,surmodel-method}
\alias{predict,surmodel-method}
\alias{predict}
\title{Predictor surogate model}
\usage{
\S4method{predict}{surmodel}(object, newdata = NULL)
}
\arguments{
\item{object}{An object of class surmodel}

\item{newdata}{An opttional vector, matrix or data.frame containing the
points where to perfom predictions. If not provided the predicted optima
will be outputed.}
}
\description{
This functions performs predictions for an surogate model.
}
