% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.SEL}
\alias{.SEL}
\alias{.X}
\alias{.X.data.frame}
\alias{.X.list}
\alias{.X.default}
\alias{.Y}
\alias{.Y.data.frame}
\alias{.Y.list}
\alias{.Y.default}
\alias{.G}
\alias{.G.data.frame}
\alias{.G.list}
\alias{.G.default}
\title{sel_funs: Selection Functions}
\usage{
.SEL(x, ...)

.X(x)

\method{.X}{data.frame}(x)

\method{.X}{list}(x)

\method{.X}{default}(x, ...)

.Y(x)

\method{.Y}{data.frame}(x)

\method{.Y}{list}(x)

\method{.Y}{default}(x, ...)

.G(x)

\method{.G}{data.frame}(x)

\method{.G}{list}(x)

\method{.G}{default}(x, ...)
}
\arguments{
\item{x}{input data to be subeseted}

\item{...}{aditional parameters}
}
\description{
Selection functions
}
