% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.SET}
\alias{.SET}
\alias{.Y<-}
\alias{.G<-}
\title{set_funs: Setting Functions}
\usage{
.SET(x, value)

.Y(x) <- value

.G(x) <- value
}
\arguments{
\item{x}{input data to be subeseted}

\item{value}{value to be set to x}

\item{...}{aditional parameters}
}
\description{
Seting functions
}
