% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{show.surmodel}
\alias{show.surmodel}
\title{Custom method for showing model objects}
\usage{
show.surmodel(object)
}
\arguments{
\item{object}{\code{surmodel} object}
}
\description{
Custom method for showing model objects
}
