% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suropt.R
\docType{package}
\name{suropt}
\alias{suropt}
\alias{suropt-package}
\title{suropt: Surrogate-based optimization}
\description{
suropt is a implementation of some surrogate-based single and
multi-objective optimization schmes.
}
\details{
The idea, is to alleviate the computational cost of optimization
exploiting meta-models of the functions.
}
