% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_hego.R
\name{train_hego}
\alias{train_hego}
\title{Trainer for a surmodel object based on the HEGO algorithm}
\usage{
train_hego(model, niter, optimizer = "gen")
}
\arguments{
\item{model}{surmodel object to be trained}

\item{niter}{integer indicating number of iterations}

\item{optimizer}{string one of: "gen" (genetic, default),
"sa" (simulated annealing) or "ps" (particle swarm)}
}
\description{
Trainer for a surmodel object based on the HEGO algorithm
}
\examples{
fn <- binh
# model <- build_surmodel(fn, 5, 2) \%>\% train_hego(1, 'sa')
}
