% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sme.R
\name{train_sme}
\alias{train_sme}
\title{Trainer for a surmodel object based on the SME algorithm}
\usage{
train_sme(model, niter, optimizer = "nsga2")
}
\arguments{
\item{model}{surmodel object to be trained}

\item{niter}{integer indicating number of iterations}

\item{optimizer}{character, only working for nsga2 by now}
}
\description{
Trainer for a surmodel object based on the SME algorithm
}
\examples{
# fn <- shaffer2
# model <- build_surmodel(fn, 10, 1) \%>\% train_sme(5)

# fn <- binh
# model <- build_surmodel(fn, 10, 2) \%>\% train_sme(5)
# suropt:::plot_predict(model)

# data <- data.frame(X.1 = runif(5), X.2 = runif(5), Y.1 = runif(5), G.1 = rnorm(5))
# x_star <- build_surmodel(data) \%>\% train_sme(-1)
}
