% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{surmodel-class}
\alias{surmodel-class}
\alias{show,surmodel-method}
\title{The \code{surmodel} class}
\usage{
\S4method{show}{surmodel}(object)
}
\arguments{
\item{object}{\code{surmodel} object}
}
\description{
The \code{surmodel} class
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show method
}}

\section{Slots}{

\describe{
\item{\code{data}}{a data.frame with the model data}

\item{\code{sur}}{a list with the surogates}

\item{\code{fn}}{a function used as high-fidelity model}
}}

