% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{surmodel-class}
\alias{surmodel-class}
\title{The \code{surmodel} class}
\description{
The \code{surmodel} class
}
\section{Slots}{

\describe{
\item{\code{data}}{bla}

\item{\code{sur}}{bla}

\item{\code{fn}}{bla}
}}

