% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sme.R
\name{train_sme}
\alias{train_sme}
\title{Trainer for a surmodel object based on the SME algorithm}
\usage{
train_sme(model, niter)
}
\arguments{
\item{model}{surmodel object to be trained}

\item{niter}{integer indicating number of iterations}
}
\description{
Trainer for a surmodel object based on the SME algorithm
}
\examples{
fn <- shaffer2
model <- build_surmodel(fn, 20, 1) \%>\% train_sme(5)

fn <- binh
model <- build_surmodel(fn, 20, 2) \%>\% train_sme(5)
}
